Add Previously Created Code
Add code created in PyCharm several weeks ago. This code requires piexif, an EXIF processing library, and Python Image Library (PIL), an image processing library to run. The code won't immediately work because the above libraries must be imported and the file path to the folder of images needs to be changed. It'd be a little weird if someone else had the exact same file path I did...

Dependencies:
 - Python Image Library (PIL)
 - piexif